feat: update fetch model notes

+++ fetch-model.sh
@@ -6,5 +6,9 @@ set -o pipefail
 
 # ONNX densenet
 mkdir -p model_repository/densenet_onnx/1
-wget -O model_repository/densenet_onnx/1/model.onnx \
-	https://github.com/onnx/models/raw/main/validated/vision/classification/densenet-121/model/densenet-7.onnx
+
+echo "https://github.com/onnx/models/raw/main/validated/vision/classification/densenet-121/model/densenet-7.onnx"
+if [ ! -f model_repository/densenet_onnx/1/model.onnx ]; then
+  wget -O model_repository/densenet_onnx/1/model.onnx \
+    https://github.com/onnx/models/raw/main/validated/vision/classification/densenet-121/model/densenet-7.onnx
+fi
